+++ packages/yaatt-cli/example-config.yaatt.json
